Move general rule to base plugin
