+++ subprojects/platform-jni/src/main/java/dev/nokee/platform/jni/internal/JavaNativeInterfaceLibraryVariantRegistrationFactory.java
@@ -168,218 +168,6 @@ public final class JavaNativeInterfaceLibraryVariantRegistrationFactory {
 					entity.addComponent(new ModelBackedNativeIncomingDependencies(path.get(), project.getObjects(), project.getProviders(), project.getExtensions().getByType(ModelLookup.class), s -> "native" + capitalize(s)));
 				}
 			}))
-			.action(ModelActionWithInputs.of(ModelComponentReference.of(IdentifierComponent.class), ModelComponentReference.of(IsVariant.class), ModelComponentReference.of(ModelState.IsAtLeastRegistered.class), ModelComponentReference.of(ModelPathComponent.class), (entity, id, tag, state, path) -> {
-				if (id.get().equals(identifier)) {
-					val registry = project.getExtensions().getByType(ModelRegistry.class);
-
-					if (!((VariantIdentifier<?>) id.get()).getUnambiguousName().isEmpty()) {
-						entity.addComponent(MultiVariantTag.tag());
-					}
-
-					val bucketFactory = new DeclarableDependencyBucketRegistrationFactory(NamedDomainObjectRegistry.of(project.getConfigurations()), new FrameworkAwareDependencyBucketFactory(project.getObjects(), new DefaultDependencyBucketFactory(NamedDomainObjectRegistry.of(project.getConfigurations()), DependencyFactory.forProject(project))));
-					val implementation = registry.register(bucketFactory.create(DependencyBucketIdentifier.of(declarable("nativeImplementation"), identifier)));
-					val linkOnly = registry.register(bucketFactory.create(DependencyBucketIdentifier.of(declarable("nativeLinkOnly"), identifier)));
-					val runtimeOnly = registry.register(bucketFactory.create(DependencyBucketIdentifier.of(declarable("nativeRuntimeOnly"), identifier)));
-
-					val sharedLibrary = registry.register(ModelRegistration.builder().mergeFrom(project.getExtensions().getByType(SharedLibraryBinaryRegistrationFactory.class).create(BinaryIdentifier.of(identifier, BinaryIdentity.ofMain("sharedLibrary", "shared library binary")))).withComponent(ExcludeFromQualifyingNameTag.tag()).withComponent(new BuildVariantComponent(identifier.getBuildVariant())).build());
-					project.getExtensions().getByType(ModelConfigurer.class).configure(ModelActionWithInputs.of(ModelComponentReference.of(IdentifierComponent.class), ModelComponentReference.of(IsBinary.class), ModelComponentReference.of(LinkLibrariesConfiguration.class), (e, idd, t, linkLibraries) -> {
-						if (idd.get().equals(ModelNodes.of(sharedLibrary).get(IdentifierComponent.class).get())) {
-							linkLibraries.configure(configureExtendsFrom(implementation.as(Configuration.class), linkOnly.as(Configuration.class)));
-						}
-					}));
-					project.getExtensions().getByType(ModelConfigurer.class).configure(ModelActionWithInputs.of(ModelComponentReference.of(IdentifierComponent.class), ModelComponentReference.of(IsBinary.class), ModelComponentReference.of(RuntimeLibrariesConfiguration.class), (e, idd, t, runtimeLibraries) -> {
-						if (idd.get().equals(ModelNodes.of(sharedLibrary).get(IdentifierComponent.class).get())) {
-							runtimeLibraries.configure(configureExtendsFrom(implementation.as(Configuration.class), runtimeOnly.as(Configuration.class)));
-						}
-					}));
-					val sharedLibraryTask = registry.register(project.getExtensions().getByType(TaskRegistrationFactory.class).create(TaskIdentifier.of(identifier, "sharedLibrary"), Task.class).build());
-					sharedLibraryTask.configure(Task.class, configureBuildGroup());
-					sharedLibraryTask.configure(Task.class, configureDescription("Assembles the shared library binary of %s.", identifier.getDisplayName()));
-					sharedLibraryTask.configure(Task.class, configureDependsOn(sharedLibrary.as(SharedLibraryBinary.class)));
-
-					val developmentBinaryProperty = registry.register(project.getExtensions().getByType(ModelPropertyRegistrationFactory.class).createProperty(ModelPropertyIdentifier.of(identifier, "developmentBinary"), Binary.class));
-					((ModelProperty<Binary>) developmentBinaryProperty).asProperty(property(of(Binary.class))).convention(project.provider(() -> new Object()).flatMap(it -> ModelElements.of(entity).element("jniJar", JniJarBinary.class).asProvider()));
-
-					project.getPlugins().withType(NativeLanguagePlugin.class, new Action<NativeLanguagePlugin>() {
-						private ModelElement compileOnly = null;
-
-						@Override
-						public void execute(NativeLanguagePlugin appliedPlugin) {
-							if (compileOnly == null) {
-								compileOnly = registry.register(bucketFactory.create(DependencyBucketIdentifier.of(declarable("nativeCompileOnly"), identifier)));
-							}
-							val sourceSet = registry.register(project.getExtensions().getByType(appliedPlugin.getRegistrationFactoryType()).create(identifier));
-							val sourceSetIdentifier = (LanguageSourceSetIdentifier) ModelNodes.of(sourceSet).get(IdentifierComponent.class).get();
-							val configurer = project.getExtensions().getByType(ModelConfigurer.class);
-							configurer.configure(ModelActionWithInputs.of(ModelComponentReference.of(IdentifierComponent.class), ModelComponentReference.of(IsDependencyBucket.class), ModelComponentReference.ofProjection(Configuration.class), (e, i, tag, configuration) -> {
-								if (((DependencyBucketIdentifier) i.get()).getOwnerIdentifier().equals(sourceSetIdentifier)) {
-									registry.instantiate(ModelAction.configure(e.getId(), Configuration.class, configureExtendsFrom(implementation.as(Configuration.class), compileOnly.as(Configuration.class))));
-									registry.instantiate(ModelAction.configure(e.getId(), Configuration.class, ConfigurationUtilsEx.configureIncomingAttributes((BuildVariantInternal) identifier.getBuildVariant(), project.getObjects())));
-									registry.instantiate(ModelAction.configure(e.getId(), Configuration.class, ConfigurationUtilsEx::configureAsGradleDebugCompatible));
-								}
-							}));
-							configurer.configure(ModelActionWithInputs.of(ModelComponentReference.of(IdentifierComponent.class), ModelComponentReference.of(IsTask.class), ModelComponentReference.ofProjection(TaskProvider.class), (e, i, t, p) -> {
-								if (((TaskIdentifier<?>) i.get()).getOwnerIdentifier().equals(sourceSetIdentifier)) {
-									e.addComponent(CompileTaskTag.tag());
-									NamedDomainObjectProvider<Task> compileTask = p.get(of(NamedDomainObjectProvider.class));
-									compileTask.configure(configureTargetPlatform(set(fromBuildVariant(identifier.getBuildVariant()))));
-								}
-							}));
-						}
-					});
-
-					val objectsTask = registry.register(project.getExtensions().getByType(TaskRegistrationFactory.class).create(TaskIdentifier.of(identifier, "objects"), Task.class).build());
-					objectsTask.configure(Task.class, configureDependsOn(sharedLibrary.as(SharedLibraryBinary.class).flatMap(binary -> binary.getCompileTasks().filter(it -> it instanceof HasObjectFiles))));
-					objectsTask.configure(Task.class, configureBuildGroup());
-					objectsTask.configure(Task.class, configureDescription("Assembles the object files of %s.", identifier.getDisplayName()));
-
-					val assembleTask = registry.register(project.getExtensions().getByType(TaskRegistrationFactory.class).create(TaskIdentifier.of(TaskName.of(ASSEMBLE_TASK_NAME), identifier), Task.class).build());
-					assembleTask.configure(Task.class, task -> {
-						task.setGroup(LifecycleBasePlugin.BUILD_GROUP);
-						if (task.getDescription() == null) {
-							task.setDescription(String.format("Assembles the outputs of %s.", identifier.getDisplayName()));
-						}
-					});
-					entity.addComponent(new AssembleTask(ModelNodes.of(assembleTask)));
-
-					val nativeRuntimeFiles = registry.register(project.getExtensions().getByType(ModelPropertyRegistrationFactory.class).createFileCollectionProperty(ModelPropertyIdentifier.of(identifier, "nativeRuntimeFiles")));
-					((ModelProperty<Set<File>>) nativeRuntimeFiles).asProperty(of(ConfigurableFileCollection.class)).from(sharedLibrary.as(SharedLibraryBinary.class).flatMap(SharedLibraryBinary::getLinkTask).flatMap(LinkSharedLibrary::getLinkedFile));
-					((ModelProperty<Set<File>>) nativeRuntimeFiles).asProperty(of(ConfigurableFileCollection.class)).from((Callable<Object>) () -> ModelNodes.of(sharedLibrary).get(DependentRuntimeLibraries.class));
-
-					ModelProperties.getProperty(sharedLibrary, "baseName").asProperty(property(of(String.class))).convention(project.provider(() -> new Object()).flatMap(it -> ModelProperties.getProperty(entity, "baseName").as(String.class).asProvider()));
-
-					val resourcePathProperty = registry.register(project.getExtensions().getByType(ModelPropertyRegistrationFactory.class).createProperty(ModelPropertyIdentifier.of(identifier, "resourcePath"), String.class));
-					((ModelProperty<String>) resourcePathProperty).asProperty(property(of(String.class))).convention(identifier.getAmbiguousDimensions().getAsKebabCase().orElse(""));
-
-					sharedLibrary.configure(SharedLibraryBinary.class, binary -> binary.getBaseName().convention(project.provider(() -> new Object()).flatMap(it -> ModelProperties.getProperty(entity, "baseName").as(String.class).asProvider())));
-
-					entity.addComponent(createdUsing(of(JniLibraryInternal.class), () -> project.getObjects().newInstance(JniLibraryInternal.class, identifier, project.getObjects())));
-
-
-					project.getExtensions().getByType(ModelConfigurer.class).configure(ModelActionWithInputs.of(ModelComponentReference.of(IdentifierComponent.class), ModelComponentReference.of(IsVariant.class), ModelComponentReference.of(ModelState.IsAtLeastRealized.class), ModelComponentReference.of(ModelPathComponent.class), ModelComponentReference.ofProjection(JniLibrary.class), (e, i, t, s, p, library) -> {
-						if (i.get().equals(identifier)) {
-							val unbuildableMainComponentLogger = new WarnUnbuildableLogger(identifier.getComponentIdentifier());
-
-							ModelNodeUtils.get(e, JniLibrary.class).getJavaNativeInterfaceJar().getJarTask().configure(configureJarTaskUsing(project.provider(() -> ModelNodeUtils.get(e, JniLibrary.class)), unbuildableMainComponentLogger));
-						}
-					}));
-
-					registry.instantiate(configureMatching(ownedBy(entity.getId()).and(subtypeOf(of(Configuration.class))), new ExtendsFromParentConfigurationAction(project, path.get())));
-				}
-			}))
 			.build();
 	}
-
-	// NOTE(daniel): I added the diagnostic because I lost about 2 hours debugging missing files from the generated JAR file.
-	//  The concept of diagnostic is something I want to push forward throughout Nokee to inform the user of possible configuration problems.
-	//  I'm still hesitant at this stage to throw an exception vs warning in the console.
-	//  Some of the concept here should be shared with the incompatible plugin usage (and vice-versa).
-	private static class MissingFileDiagnostic {
-		private boolean hasAlreadyRan = false;
-		private final List<File> missingFiles = new ArrayList<>();
-
-		public void logTo(Logger logger) {
-			if (!missingFiles.isEmpty()) {
-				StringBuilder builder = new StringBuilder();
-				builder.append("The following file");
-				if (missingFiles.size() > 1) {
-					builder.append("s are");
-				} else {
-					builder.append(" is");
-				}
-				builder.append(" missing and will be absent from the JAR file:").append(System.lineSeparator());
-				for (File file : missingFiles) {
-					builder.append(" * ").append(file.getPath()).append(System.lineSeparator());
-				}
-				builder.append("We recommend taking the following actions:").append(System.lineSeparator());
-				builder.append(" - Verify 'nativeRuntimeFile' property configuration for each variants").append(System.lineSeparator());
-				builder.append("Missing files from the JAR file can lead to runtime errors such as 'NoClassDefFoundError'.");
-				logger.warn(builder.toString());
-			}
-		}
-
-		public void missingFiles(List<File> missingFiles) {
-			this.missingFiles.addAll(missingFiles);
-		}
-
-		public void run(Consumer<MissingFileDiagnostic> action) {
-			if (!hasAlreadyRan) {
-				action.accept(this);
-				hasAlreadyRan = false;
-			}
-		}
-	}
-
-	private static Action<Jar> configureJarTaskUsing(Provider<JniLibrary> library, WarnUnbuildableLogger logger) {
-		val runnableLogger = onlyOnce(logger::warn);
-		return task -> {
-			MissingFileDiagnostic diagnostic = new MissingFileDiagnostic();
-			task.doFirst(new Action<Task>() {
-				@Override
-				public void execute(Task task) {
-					diagnostic.run(warnAboutMissingFiles(task.getInputs().getSourceFiles()));
-					diagnostic.logTo(task.getLogger());
-				}
-
-				private Consumer<MissingFileDiagnostic> warnAboutMissingFiles(Iterable<File> files) {
-					return diagnostic -> {
-						ImmutableList.Builder<File> builder = ImmutableList.builder();
-						File linkedFile = library.map(it -> it.getSharedLibrary().getLinkTask().get().getLinkedFile().get().getAsFile()).get();
-						for (File file : files) {
-							if (!file.exists() && !file.equals(linkedFile)) {
-								builder.add(file);
-							}
-						}
-						diagnostic.missingFiles(builder.build());
-					};
-				}
-			});
-			task.from(library.flatMap(it -> {
-				// TODO: The following is debt that we accumulated from gradle/gradle.
-				//  The real condition to check is, do we know of a way to build the target machine on the current host.
-				//  If yes, we crash the build by attaching the native file which will tell the user how to install the right tools.
-				//  If no, we can "silently" ignore the build by saying you can't build on this machine.
-				//  One consideration is to deactivate publishing so we don't publish a half built jar.
-				//  TL;DR:
-				//    - Single variant where no toolchain could ever build the binary (unavailable) => print a warning
-				//    - Single variant where no toolchain is found to build the binary (unbuildable) => fail
-				//    - Single variant where toolchain is found to build the binary (buildable) => build (and hopefully succeed)
-				if (task.getName().equals("jar")) {
-					// TODO: Test this scenario in a functional test
-					if (it.getSharedLibrary().isBuildable()) {
-						return it.getNativeRuntimeFiles().getElements();
-					} else {
-						runnableLogger.run();
-						return ProviderUtils.fixed(emptyList());
-					}
-				}
-				return it.getNativeRuntimeFiles().getElements();
-			}), spec -> {
-				// Don't resolve the resourcePath now as the JVM Kotlin plugin (as of 1.3.72) was resolving the `jar` task early.
-				spec.into(library.map(JniLibrary::getResourcePath));
-			});
-		};
-	}
-
-	//region Target platform
-	public static <SELF extends Task> Action<SELF> configureTargetPlatform(BiConsumer<? super SELF, ? super PropertyUtils.Property<? extends NativePlatform>> action) {
-		return task -> {
-			action.accept(task, wrap(targetPlatformProperty(task)));
-		};
-	}
-
-	private static NativePlatform fromBuildVariant(BuildVariant buildVariant) {
-		return NativePlatformFactory.create(buildVariant).get();
-	}
-
-	private static Property<NativePlatform> targetPlatformProperty(Task task) {
-		if (task instanceof AbstractNativeCompileTask) {
-			return ((AbstractNativeCompileTask) task).getTargetPlatform();
-		} else if (task instanceof SwiftCompile) {
-			return ((SwiftCompile) task).getTargetPlatform();
-		} else {
-			throw new IllegalArgumentException();
-		}
-	}
-	//endregion
 }
